requirements.txt
fastapi
uvicorn
streamlit
pydantic
httpx
aiohttp
python-dotenv
pyyaml
arxiv
beautifulsoup4
lxml
markdown
openai
tenacity
